refactor: add directory structure for history
