+++ src/index.js
@@ -1,6 +1,6 @@
 import { initEventListener } from "./util/initEventListener.js";
-import { HeaderView } from "./view/HeaderView.js";
-import { HistoryView } from "./view/HistoryView.js";
+import HeaderView from "./view/Header/index.js";
+import HistoryView from "./view/History";
 import MainView from "./view/Main/index.js";
 
 (function init() {
