+++ src/view/Header/index.js
@@ -1,4 +1,4 @@
-export const HeaderView = () => {
+const HeaderView = () => {
   return ` 
     <header class="header">
         <h1 class="header__title">TODO LIST</h1>
@@ -7,3 +7,5 @@ export const HeaderView = () => {
     </header>
     `;
 };
+
+export default HeaderView;
