+++ src/view/History/HistoryListView.js
@@ -1,0 +1,10 @@
+import { historyDataList } from "../../model/model.js";
+import HistoryCardView from "./Card/index.js";
+
+export const HistoryListView = () => {
+  return `
+      <ul class="history__list">
+        ${historyDataList.map((history) => HistoryCardView(history)).join("")}
+      </ul>
+    `;
+};
